Examples Jasmine spec files
Included are example Jasmine spec files for testing the JavaScript file
src/SimpleDemo.js file which is used in the supplied Demo.html file.
The example Jasmine files show a simple spec, use of some matchers and
use of not with matchers.  there is an example custom matcher and an
example spy.